PR 34677: Add simple pipeline benchmark app
This change adds a simple benchmark application that can help us assess the performance of our pipeline

The benchmark uses one event source as an input. The pipeline uses one metadata filter and two outputs,
each with its own simple filter. Current throughput on my dev machine is about 2000 events/second.
The benchmark app is not using CPU efficiently--one core is at 100% where the rest sit idle.
We have some work to do :-)

In addition, I changed a few things about the core implementation:
-----------------------------------------------------------------
1. The filter method is called "Evaluate" instead of "Filter" and returns a FilterResult enumeration.
This is to make it clearer what the result of a filter is.
2. DiagnosticsPipeline Inputs and Sinks are now IEnumerable<something> instead of IReadOnlyCollection<something>
It makes it easier to work with them programmatically.
3. I implemented the drop filter. Ended up not using it in the benchmark, but since it was part of v1
plan, I included it in this PR.
4. Fixed a bug in the EventData cloning code
5. Moved the IncludeCondition of the metadata filter from the metadata class to the filter itself.
It does not feel right to have it on the metadata.

Next step: do the DataFlow conversion and see how much we improve or regress :-)