Improve batch timer handling in DiagnosticPipeline (#325)
* Improve batch timer handling in DiagnosticPipeline

This is an attempt to reduce the possibility of ending in a state described in issue 311 https://github.com/Azure/diagnostics-eventflow/issues/311 It is not entirely clear to me what the problem is there because I have a very hard time reproducing the issue, but the one time when it did repro with debugger attached, I saw that the batcher timer was not being fired. As far as I could tell there was nothing wrong with EventFlow pipeline.

I suspect this is an initialization timing issue that results in thread pool thred exhaustion and the timer proc not being scheduled as it should. One can get more or less to the same state when threadpool threads are manually suspended (the timer is no loger fired, and no new threads are added to the thread pool).

This change ensures that only one timer proc is ever running, so it should help and in general is a good rule to follow.